Reorganizes tests betweeen acceptance/integration.  Ape project is now /tests/acceptnace.
