+++ tests/acceptance/agents/test_coordinator_agent.py
