--- tests/acceptance/blockchain/actors/conftest.py
@@ -1,1 +1,0 @@
-
