--- tests/acceptance/characters/conftest.py
@@ -1,1 +1,0 @@
-
