+++ tests/acceptance/contracts/ConditionNFT.sol
