+++ tests/acceptance/pytest.ini
@@ -1,0 +1,7 @@
+[pytest]
+addopts = -s --maxfail 1
+testpaths =
+    actors/
+    agents/
+    conditions/
+    interfaces/
