+++ tests/fixtures.py
@@ -17,6 +17,7 @@ from web3 import Web3
 import nucypher
 import tests
 from nucypher.blockchain.economics import Economics
+from nucypher.blockchain.eth.clients import EthereumClient
 from nucypher.blockchain.eth.interfaces import BlockchainInterfaceFactory
 from nucypher.blockchain.eth.registry import (
     LocalContractRegistry,
@@ -472,6 +473,14 @@ def highperf_mocked_bob(fleet_of_highperf_mocked_ursulas):
 #
 # CLI
 #
+
+@pytest.fixture(scope="function")
+def mock_funding_and_bonding(testerchain, mocker):
+    mocker.patch(
+        "nucypher.blockchain.eth.actors.Operator.get_staking_provider_address",
+        return_value=testerchain.stake_providers_accounts[0],
+    )
+    mocker.patch.object(EthereumClient, "get_balance", return_value=1)
 
 
 @pytest.fixture(scope='function')
