+++ tests/integration/characters/test_freerider_attacks.py
