+++ tests/integration/cli/test_ursula_cli_ip_detection.py
@@ -1,4 +1,5 @@
 from nucypher.blockchain.eth.actors import Operator
+from nucypher.blockchain.eth.trackers.dkg import ActiveRitualTracker
 from nucypher.cli.commands import ursula
 from nucypher.cli.main import nucypher_cli
 from nucypher.config.characters import UrsulaConfiguration
@@ -79,6 +80,9 @@ def test_ursula_run_ip_checkup(
     ursula_test_config,
     tempfile_path,
 ):
+
+    # Mock DKG
+    mocker.patch.object(ActiveRitualTracker, 'start', autospec=True)
 
     # Mock IP determination
     target = 'nucypher.cli.actions.configure.determine_external_ip_address'
