+++ tests/integration/network/test_availability.py
